requirements.txt
aiofiles
gallery-dl
pillow
py-cord
python-dotenv
python-ffmpeg
ffmpeg-asyncio
yt-dlp